--- data/dbConfig.js
@@ -1,4 +1,0 @@
-const knex = require('knex');
-const knexConfig = require('../knexfile.js');
-
-module.exports = knex(knexConfig.development);
