--- server.js
@@ -1,15 +1,0 @@
-const express = 'express';
-
-const server = express();
-
-server.get('/', (req, res) => {
-  res.send(`<h2>Let's write some middleware!</h2>`)
-});
-
-//custom middleware
-
-function logger(req, res, next) {
-
-};
-
-module.exports = server;
